test(api): split azure exchange tests out and fix a misplaced doc comment
Addresses two CodeRabbit nitpicks on the Azure RI exchange tests.

The doc comment describing TestCheckAzureExchangeMoneyGuardrails_NilPreviewRefused
sat above TestCheckAzureExchangeMoneyGuardrails_CurrencyCaseInsensitive, so
both comments stacked over the wrong test and the nil-preview rationale
appeared to explain the currency rule. Move it to directly precede the test
it documents; neither test body changes.

Move the Azure compatible-offerings, execute, money-guardrail and
subscription-scoping suites into handler_ri_exchange_azure_test.go in the
same package. handler_ri_exchange_test.go had grown to 2788 lines against
the project's 500-line guideline, and the Azure block was a self-contained
~1000-line section with its own mocks and helpers. Same package, so shared
setup keeps working unchanged.

No test was added, removed or altered: 86 test functions before, 86 after.
